feature: integrated main cart component with Cart_Songs_View component, fixed unused vars and cleaned up imports, fixed seed file to associate mixtapes with order IDs, removed use of local storage
