+++ client/components/index.js
@@ -9,6 +9,4 @@ export {default as UserHome} from './user-home'
 export {Login, Signup} from './auth-form'
 export {default as Single_Song} from './single-song'
 export {default as Cart} from './cart/index'
-export {default as Cart_Songs_View} from './cart-songs-view'
-
-
+// export {default as Cart_Songs_View} from './cart-songs-view'
