+++ script/seed.js
@@ -60,6 +60,11 @@ async function seed() {
   await allOrders[5].setUser(allUsers[0])
   await allOrders[6].setUser(allUsers[3])
 
+  await allMixtapes[0].setOrder(allOrders[6])
+  await allMixtapes[1].setOrder(allOrders[3])
+  await allMixtapes[2].setOrder(allOrders[2])
+  await allMixtapes[5].setOrder(allOrders[0])
+
   await allMixtapes[0].addSong(allSongs[3])
   await allMixtapes[0].addSong(allSongs[6])
   await allMixtapes[1].addSong(allSongs[(9, 0)])
